% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_frames.R
\name{as_snowdl_dates}
\alias{as_snowdl_dates}
\title{Create \code{snowdl_dates} from a vector}
\usage{
as_snowdl_dates(x)
}
\arguments{
\item{x}{\verb{[Date]} The vector of dates for which to download data.}
}
\description{
Creates an object of class \code{snowdl_dates} from a vector of desired dates
}
\details{
This coerces a user-specified vector of dates to a \code{snowdl_dates}
\code{data.frame}. For more options, see
}
\examples{

\dontrun{

my_seq <- seq(as.Date("2020-01-01"), as.Date("2020-01-19"), by = "3 days")
as_snowdl_dates(my_seq)

}

}
