% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crop_or_reproject}
\alias{crop_or_reproject}
\title{Crop or Reproject a Raster}
\usage{
crop_or_reproject(r, crop, reproject, method)
}
\arguments{
\item{r}{The RasterLayer to crop or reproject.}

\item{crop}{A Raster* or Extent to crop to.}

\item{reproject}{A Raster* to reproject to.}

\item{method}{The method if reprojecting.}
}
\description{
Decides whether to crop or reproject based on arguments.
}
\details{
Meant for internal use.
}
