% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{dates_SNODAS}
\alias{dates_SNODAS}
\title{Process \code{snowdl_dates} for SNODAS}
\usage{
dates_SNODAS(dates, masked = TRUE)
}
\arguments{
\item{dates}{\verb{[snowdl_dates]} The object that defines the dates for which to
download data. See ?snowdl_dates.}

\item{masked}{\verb{[logical = TRUE]} Determines whether the downloaded data will
be masked or unmasked. See details.}
}
\value{
An object of class \code{SNODAS_dates} to be passed to
\code{download_SNODAS()}.
}
\description{
Process \code{snowdl_dates} object for use with SNODAS
}
\details{
This function determines URLs to download SNODAS data over HTTP from
this site: "https://noaadata.apps.nsidc.org/NOAA/G02158/".

On the argument \code{masked}: SNODAS makes predictions slightly outside of the
contiguous United States (unmasked), but crops the predictions to the
contiguous US (masked). Note that while masked rasters are available
beginning September 2003, unmasked rasters are only available beginning
in December 2009.
}
\examples{

\dontrun{

dd <- snowdates_ydoy(2008:2010, 1:3, all_combos = TRUE)
dates_SNODAS(dd, masked = TRUE)
dates_SNODAS(dd, masked = FALSE)

}

}
