% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{download_SNODAS}
\alias{download_SNODAS}
\title{Download SNODAS}
\usage{
download_SNODAS(date, out_dir = ".", overwrite = FALSE)
}
\arguments{
\item{date}{\verb{[date, POSIXt]} The (1) date for which to download data.}

\item{out_dir}{\verb{[character = "."]} The directory in which to save downloaded
file.}

\item{overwrite}{\verb{[logical = FALSE]} If data for \code{date} exists in \code{out_dir},
should it be overwritten?}
}
\description{
Downloads SNODAS data for a single date
}
\details{
This function downloads SNODAS data over FTP from
this site: "ftp://sidads.colorado.edu/DATASETS/NOAA/G02158/masked/".

This function is called by wrapper function \strong{TBD}.
}
\examples{
\dontrun{
download_SNODAS(as.Date("2020-02-02"), out_dir = tempdir())
}
}
