% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{doy}
\alias{doy}
\title{Return day-of-year for a date}
\usage{
doy(d)
}
\arguments{
\item{d}{A \code{Date} object}
}
\description{
Internal function that calculates day-of-year for a date
}
\details{
Meant for internal use.
}
