% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMWFR_fun.R
\name{e_key}
\alias{e_key}
\title{Set ECMWFR key}
\usage{
e_key(...)
}
\arguments{
\item{...}{All arguments passed to \code{ecmwfr::wf_set_key()}}
}
\description{
Wrapper function to set user's ECMWFR key
}
\examples{

\dontrun{

e_key(user = "test@mail.com", key = "123", service = "cds")

}

}
\seealso{
\code{\link[ecmwfr:wf_set_key]{wf_set_key}()}
}
