% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMWF_fun.R
\name{e_request}
\alias{e_request}
\title{Create ECMWFR Request}
\usage{
e_request(
  variable = c("snow_density", "snow_depth", "snow_depth_water_equivalent"),
  years,
  months,
  days,
  time = "06:00",
  area = c(49, -125, 24.5, -66.5),
  out_file
)
}
\arguments{
\item{variable}{\verb{[character]} Only one of \code{"snow_density"},
\code{"snow_depth"}, \code{"snow_depth_water_equivalent"}.}

\item{years}{\verb{[character]} Years for which to download data.}

\item{months}{\verb{[character]} Months for which to download data.}

\item{days}{\verb{[character]} Days for which to download data.}

\item{time}{\verb{[character = "06:00"]} Hours for which to download data,
formatted as HH:MM. Hourly data are available for this dataset.}

\item{area}{\verb{[numeric]} Vector specifying bounding box of spatial
subset to download. Should be in lat/lon, counter-clockwise from top.
Order is c(ymax, xmin, ymin, xmax). Default downloads all of USA.}

\item{out_file}{\verb{[character]} Name of output file. Should end in ".nc".}
}
\value{
List
}
\description{
Creates a list with data request parameters
}
\details{
See vignette "CDS functionality" in package \code{ecmwfr} for
details on creating requests. Requests can also be created by visiting
web API, downloading python code, and using an RStudio function to
convert to list.
}
\examples{

req_list <- e_request(variable = "snow_density",
                      years = 2018:2019,
                      months = 3:4,
                      days = 15,
                      out_file = "test.nc")

}
\seealso{
\href{https://cran.r-project.org/web/packages/ecmwfr/vignettes/cds_vignette.html}{CDS functionality vignette}
}
