% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daymet_fun.R
\name{get_daymet_swe}
\alias{get_daymet_swe}
\title{Download Daymet SWE Data}
\usage{
get_daymet_swe(year, tile, out_dir = "data")
}
\arguments{
\item{year}{\verb{[numeric]} Year(s) for which to download data.}

\item{tile}{\verb{[numeric]} Tile(s) for which to download data. See details.}

\item{out_dir}{\verb{[character = "data"]} The directory where data will be
downloaded. It will be created if it does not exist.}
}
\value{
Returns character vector with full paths to downloaded files.
}
\description{
Downloads snow-water equivalent data from Daymet
}
\details{
This function creates a text string for the correct URL for each
year/tile combination, which will be passed to \code{utils::download.file()} to
download the data.

More details on arguments:
\itemize{
\item \code{year} can be a vector of any number of years. It will be combined with
\code{tile} to return all combinations of \code{year} and \code{tile}.
\item \code{tile} can be a vector of any number of tiles. A figure with the Daymet
tiles can be found
\href{https://daymet.ornl.gov/static/graphics/TilesV4_Daymet.png}{here}
}
}
\examples{

\dontrun{

get_daymet_swe(year = c(2018, 2019),
               tile = c(12095, 12096),
               out_dir = tempdir())
}

}
