% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMWF_fun.R
\name{get_request}
\alias{get_request}
\title{Get ECMWF Data}
\usage{
get_request(...)
}
\arguments{
\item{...}{All arguments passed to \code{ecmwfr::wf_request()}}
}
\value{
Returns character string with path to downloaded file.
}
\description{
Download ECMWF data from a request
}
\examples{

\dontrun{

e_key(user = "112233", key = "123456789", service = "cds")

req_list <- e_request(variable = "snow_density",
                      years = 2018:2019,
                      months = 3:4,
                      days = 15,
                      out_file = "test.nc")

get_request(user = "112233",
            request = req_list,
            transfer = TRUE,
            path = ".",
            verbose = TRUE)

}

}
\seealso{
\code{\link[ecmwfr:wf_request]{wf_request}()}
}
