% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{rasterize_SNODAS}
\alias{rasterize_SNODAS}
\title{Rasterize SNODAS files}
\usage{
rasterize_SNODAS(
  data_dir = "data",
  out_dir = "data/geo",
  rm_data = TRUE,
  format = c("raster", "GTiff"),
  extent = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_dir}{\verb{[character = "data"]} Directory where \verb{SNODAS_*} folders are
located}

\item{out_dir}{\verb{[character = "data"]} Directory where rasters should be
saved}

\item{rm_data}{\verb{[logical = TRUE]} Delete untarred files after saving to
raster?}

\item{format}{\verb{[character = "raster"]} Output format of raster. Either
\code{"raster"} (.grd) or \code{"GTiff"} (.tif).}

\item{extent}{\verb{[Extent = NULL]} Optional. Extent for cropping SNODAS rasters.
If \code{NULL} (default), no cropping occurs.}
}
\description{
Saves as raster already untarred files in SNODAS directory
}
\details{
Blah blah. \strong{TBD}
}
