% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMWFR_fun.R
\name{set_ecmwfr_key}
\alias{set_ecmwfr_key}
\title{Set ECMWFR key}
\usage{
set_ecmwfr_key(...)
}
\description{
Wrapper function to set user's ECMWFR key
}
\examples{

\dontrun{

set_ecmwfr_key(user = "test@mail.com", key = "123", service = "cds")

}

}
\seealso{
\code{\link[ecmwfr::wf_set_key]{wf_set_key}}()
}
