% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{SNODAS_pc}
\alias{SNODAS_pc}
\title{Display SNODAS Product Codes}
\usage{
SNODAS_pc()
}
\description{
Displays state variables available in SNODAS data
}
\examples{
SNODAS_pc()


}
\seealso{
See Table 1 of SNODAS user guide at
https://nsidc.org/data/g02158/versions/1
}
