% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_frames.R
\name{snowdates_seq}
\alias{snowdates_seq}
\title{Create sequence of \code{snowdl_dates}}
\usage{
snowdates_seq(start, end)
}
\arguments{
\item{start}{\verb{[Date]} The first date for which to download data.}

\item{end}{\verb{[Date]} The last date for which to download data.}
}
\description{
Creates an object of class \code{snowdl_dates} from a start and end date
}
\details{
This function assumes you want every date from start to end.
}
\examples{

\dontrun{

snowdates_seq(as.Date("2020-01-01"), as.Date("2020-01-15"))

}

}
