% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_frames.R
\name{snowdates_ydoy}
\alias{snowdates_ydoy}
\title{Create \code{snowdl_dates} for combinations of year and DOY}
\usage{
snowdates_ydoy(y, doy, all_combos = TRUE)
}
\arguments{
\item{y}{\verb{[numeric]} A vector of years for which to download data.}

\item{doy}{\verb{[numeric]} A vector of day-of-year (1:366) values for which to
download data.}

\item{all_combos}{\verb{[logical = TRUE]} Should \code{snowdl} create all combinations
of \code{y} and \code{doy} or just match (or recycle) each value. See details.}
}
\description{
Creates an object of class \code{snowdl_dates} from combinations of year and
day-of-year
}
\details{
This function makes it easier to get a subset of days across
various years.

The argument \code{all_combos} decides whether you are asking for
all combinations of \code{y} and \code{doy} (\code{TRUE}) or if you want to match each \code{y}
with a single \code{doy} (using R's recycling rules if necessary). I.e.,
\code{all_combos = TRUE} uses \code{expand.grid()} whereas \code{all_combos = FALSE} uses
\code{data.frame()} directly.
}
\examples{

\dontrun{

# All combinations of year and day-of-year
snowdates_ydoy(y = 2020:2021, doy = 1:6, all_combos = TRUE)

# Match vectors year and day-of-year (with recycling)
snowdates_ydoy(y = 2020:2021, doy = 1:6, all_combos = FALSE)

}

}
