% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdl-package.R
\docType{package}
\name{snowdl-package}
\alias{snowdl}
\alias{snowdl-package}
\title{snowdl: Download Snow Data}
\description{
Allows downloading and processing of snow data (e.g., depth, density, snow-water equivalent) from various sources (e.g., Daymet, SNODAS).
}
\author{
\strong{Maintainer}: Brian J. Smith \email{brian.smith@usu.edu}

}
\keyword{internal}
