% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{unpack_dir_SNODAS}
\alias{unpack_dir_SNODAS}
\title{Unpack all SNODAS tarballs}
\usage{
unpack_dir_SNODAS(tar_dir = ".", out_dir = "data", rm_tar = FALSE)
}
\arguments{
\item{tar_dir}{\verb{[character = "."]} The directory from which to unpack all
SNODAS tarballs}

\item{out_dir}{\verb{[character = "."]} The directory where outputs will be saved}

\item{rm_tar}{\verb{[logical = FALSE]} Should tarballs be deleted after unpacking?}
}
\description{
Unpacks all SNODAS tarballs in a directory
}
\details{
This function unpacks tarballs and returns SWE and snow depth files
with new names in a user-defined directory.

This is a remnant of the previous (v0.1.0) workflow that takes just a
directory as an input. It remains available for convenience, but the
current (v1.0.0) workflow tracks the files in a custom \code{data.frame} and
users should use the function TBD.
}
\examples{
\dontrun{

# Just two dates
dts <- as_snowdl_dates(as.Date("2020-02-02") + 1:2)

# Process dates
SNODAS_dts <- dates_SNODAS(dts)

# Download to temporary directory
dd <- tempdir()
res <- download_SNODAS(SNODAS_dts, out_dir = dd)

# Unpack
unpack_dir_SNODAS(dd)

}
}
