% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{unpack_SNODAS}
\alias{unpack_SNODAS}
\title{Unpack SNODAS tarballs}
\usage{
unpack_SNODAS(dates, out_dir = "data", rm_tar = FALSE)
}
\arguments{
\item{dates}{\verb{[SNODAS_download]} The object returned by \code{download_SNODAS()}}

\item{out_dir}{\verb{[character = "."]} The directory where outputs will be saved}

\item{rm_tar}{\verb{[logical = FALSE]} Should tarballs be deleted after unpacking?}
}
\description{
Unpacks SNODAS tarballs from a \code{SNODAS_download} object
}
\details{
This function unpacks tarballs and returns SWE and snow depth files
with new names in a user-defined directory.

This is the preferred unpacking function for v1.0.0.
}
\examples{
\dontrun{

# Three dates; one is unavailable
dts <- as_snowdl_dates(as.Date(c("2000-02-02", "2010-02-02", "2020-02-02")))

# Process dates
SNODAS_dts <- dates_SNODAS(dts)

# Download to temporary directory
dd <- tempdir()
dl <- download_SNODAS(SNODAS_dts, out_dir = dd)

# Unpack
res <- unpack_SNODAS(dl, out_dir = dd)

}
}
\seealso{
\code{\link[=unpack_dir_SNODAS]{unpack_dir_SNODAS()}} for an older approach (v0.1.0) that unpacks
all tarballs in a directory without a \code{SNODAS_download} object.
}
