% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{yr}
\alias{yr}
\title{Return year for a date}
\usage{
yr(d)
}
\arguments{
\item{d}{A \code{Date} object}
}
\description{
Internal function that calculates year for a date
}
\details{
Meant for internal use.
}
