% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daymet_fun.R
\name{mosaic_daymet}
\alias{mosaic_daymet}
\title{Mosaic Daymet files}
\usage{
mosaic_daymet(
  dir,
  days = 1:365,
  out_dir = ".",
  format = c("raster", "GTiff"),
  extent = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dir}{\verb{[character]} Directory where Daymet data are stored}

\item{days}{\verb{[integer = 1:365]} Integer(ish) vector giving the days you want
retained in the data. Defaults to all days (1:365). See note in Details.}

\item{out_dir}{\verb{[character = "data"]} Directory to save mosaicked output.}

\item{format}{\verb{[character = "raster"]} Output format of raster. Either
\code{"raster"} (.grd) or \code{"GTiff"} (.tif).}

\item{extent}{\verb{[Extent = NULL]} Optional. Extent for cropping SNODAS rasters.
If \code{NULL} (default), no cropping occurs.}

\item{verbose}{\verb{[logical = TRUE]} Determines whether the user will be
notified of progress.}
}
\description{
Mosaic daymet tiles by year and day
}
\details{
This function takes the tiles downloaded by
\code{\link{get_daymet_swe}()} and mosaics them together by year and day.
Each Daymet NetCDF file has 365 bands corresponding to days of the year. This
function finds all the \verb{*.nc} files for each year in the folder, then
determines which bands (days) to load with the argument \code{days}, so
processing is sped up noticeably if you only specify the days you need. The
output is then saved in a single-band GeoTIFF labeled with yyyy_day in the
filename.

A note on Daymet days. During leap years, there are still only 365 bands in
the raster. December 31st is omitted during leap years.
}
