% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{rasterize_SNODAS}
\alias{rasterize_SNODAS}
\title{Rasterize SNODAS files}
\usage{
rasterize_SNODAS(
  dates,
  out_dir = "data/geo",
  rm_data = FALSE,
  format = "GTiff",
  crop = NULL,
  reproject = NULL,
  method = "near",
  verbose = TRUE
)
}
\arguments{
\item{dates}{\verb{[SNODAS_unpacked]} Object returned by \code{unpack_SNODAS()}}

\item{out_dir}{\verb{[character = "data"]} Directory where rasters should be
saved}

\item{rm_data}{\verb{[logical = FALSE]} Delete untarred files after saving to
raster?}

\item{format}{\verb{[character = "GTiff"]} Output format of raster. Only \code{"GTiff"}
(.tif) is currenty supported.}

\item{crop}{\verb{[SpatRaster, SpatExtent = NULL]} Optional. Template for cropping
SNODAS rasters. If \code{NULL} (default), no cropping occurs. \code{reproject} takes
priority over \code{crop} if both are specified.}

\item{reproject}{\verb{[SpatRaster = NULL]} Optional. Raster for reprojecting
SNODAS rasters. If \code{NULL} (default), no reprojection occurs. \code{reproject}
takes priority over \code{crop} if both are specified.}

\item{method}{\verb{[character = "near"]} Method used to compute values if
reprojecting raster. See ?terra::project for options. Ignored if \code{reproject}
is \code{NULL}.}

\item{verbose}{\verb{[logical = TRUE]} Determines whether the user will be
notified of progress in the console.}
}
\value{
Returns a \code{data.frame} of class \code{TBD} with metadata about each
raster. As a side-effect, it also saves the rasters to disk, with one
file for each date. Each file has 3 layers, which should retain their
names, but are (1) SWE in m (\code{swe}), (2) Snow Depth in m (\code{depth}), and (3)
Snow Pack Average Temperature in K (\code{temp}).
}
\description{
Saves as raster already untarred files in SNODAS directory
}
\examples{

\dontrun{
# Three dates; one is unavailable
dts <- as_snowdl_dates(as.Date(c("2000-02-02", "2010-02-02", "2020-02-02",
                                  "2017-03-30")))

# Process dates
SNODAS_dts <- dates_SNODAS(dts)

# Download to temporary directory
dd <- tempdir()
dl <- download_SNODAS(SNODAS_dts, out_dir = dd)

# Unpack
up <- unpack_SNODAS(dl, out_dir = dd)

# Rasterize
res <- rasterize_SNODAS(up, out_dir = dd)

}

}
\seealso{
\code{\link{SNODAS_pc}}\verb{()} for the full names and units of each layer
}
