% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{rasterize_SNODAS}
\alias{rasterize_SNODAS}
\title{Rasterize SNODAS files}
\usage{
rasterize_SNODAS(
  data_dir = "data",
  out_dir = "data/geo",
  rm_data = TRUE,
  format = c("raster", "GTiff"),
  crop = NULL,
  reproject = NULL,
  method = "ngb",
  verbose = TRUE
)
}
\arguments{
\item{data_dir}{\verb{[character = "data"]} Directory where \verb{SNODAS_*} folders are
located}

\item{out_dir}{\verb{[character = "data"]} Directory where rasters should be
saved}

\item{rm_data}{\verb{[logical = TRUE]} Delete untarred files after saving to
raster?}

\item{format}{\verb{[character = "raster"]} Output format of raster. Either
\code{"raster"} (.grd) or \code{"GTiff"} (.tif).}

\item{crop}{\verb{[Extent = NULL]} Optional. Extent for cropping Daymet rasters.
If \code{NULL} (default), no cropping occurs. \code{reproject} takes priority over
\code{crop} if both are specified.}

\item{reproject}{\verb{[Raster* = NULL]} Optional. Raster for reprojecting Daymet
rasters. If \code{NULL} (default), no reprojection occurs. \code{reproject} takes
priority over \code{crop} if both are specified.}

\item{method}{\verb{[character = "ngb"]} Method used to compute values if
reprojecting raster. Either "ngb" or "bilinear". Ignored if \code{reproject} is
\code{NULL}.}

\item{verbose}{\verb{[logical = TRUE]} Determines whether the user will be
notified of progress.}
}
\description{
Saves as raster already untarred files in SNODAS directory
}
\details{
Blah blah. \strong{TBD}
}
