% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNODAS_fun.R
\name{unpack_SNODAS}
\alias{unpack_SNODAS}
\title{Unpack SNODAS tarballs}
\usage{
unpack_SNODAS(tar_dir = ".", new_dir = "data", rm_tar = TRUE)
}
\arguments{
\item{tar_dir}{\verb{[character = "."]} The directory from which to unpack all
SNODAS tarballs}

\item{new_dir}{\verb{[character = "."]} The directory where outputs will be saved}

\item{rm_tar}{\verb{[logical = TRUE]} Should tarball be deleted after unpacking?}
}
\description{
Unpacks all SNODAS tarballs in a directory
}
\details{
This function unpacks tarballs and returns SWE and snow depth files
with new names in a user-defined directory.

Will probably end up getting called by a wrappper. \strong{TBD}
}
\examples{
\dontrun{

dd <- tempdir()

download_SNODAS(as.Date("2020-02-02"), out_dir = dd)
unpack_SNODAS(dd)

}
}
